Refactor device.py into a few separate files in device directory.
Preserve the interface using the device/__init__.py file.

Clean up or hide all pylint errors.

Tags, Device, and Attributes have dependencies on each other, so
might as well remain in the same file.
