+++ pyudev/device/_device.py
@@ -17,128 +17,39 @@
 
 
 """
-    pyudev.device
-    =============
+    pyudev.device._device
+    =====================
 
     Device class implementation of :mod:`pyudev`.
 
     .. moduleauthor::  Sebastian Wiesner  <lunaryorn@gmail.com>
 """
 
-from __future__ import (print_function, division, unicode_literals,
-                        absolute_import)
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
 
 import os
-from collections import Mapping, Container, Iterable
+from collections import Container
+from collections import Iterable
+from collections import Mapping
 from datetime import timedelta
 
-from pyudev._util import (ensure_byte_string, ensure_unicode_string,
-                          udev_list_iterate, string_to_bool,
-                          get_device_type)
+from pyudev.device._errors import DeviceNotFoundAtPathError
+from pyudev.device._errors import DeviceNotFoundByNameError
+from pyudev.device._errors import DeviceNotFoundByNumberError
+from pyudev.device._errors import DeviceNotFoundInEnvironmentError
+from pyudev._util import ensure_byte_string
+from pyudev._util import ensure_unicode_string
+from pyudev._util import get_device_type
+from pyudev._util import string_to_bool
+from pyudev._util import udev_list_iterate
 
-
-__all__ = ['Device', 'Attributes', 'Tags',
-           'DeviceNotFoundError', 'DeviceNotFoundAtPathError',
-           'DeviceNotFoundByNameError', 'DeviceNotFoundByNumberError',
-           'DeviceNotFoundInEnvironmentError']
-
-
-class DeviceNotFoundError(LookupError):
-    """
-    An exception indicating that no :class:`Device` was found.
-
-    .. versionchanged:: 0.5
-       Rename from ``NoSuchDeviceError`` to its current name.
-    """
-
-
-class DeviceNotFoundAtPathError(DeviceNotFoundError):
-    """
-    A :exc:`DeviceNotFoundError` indicating that no :class:`Device` was
-    found at a given path.
-    """
-
-    def __init__(self, sys_path):
-        DeviceNotFoundError.__init__(self, sys_path)
-
-    @property
-    def sys_path(self):
-        """
-        The path that caused this error as string.
-        """
-        return self.args[0]
-
-    def __str__(self):
-        return 'No device at {0!r}'.format(self.sys_path)
-
-
-class DeviceNotFoundByNameError(DeviceNotFoundError):
-    """
-    A :exc:`DeviceNotFoundError` indicating that no :class:`Device` was
-    found with a given name.
-    """
-
-    def __init__(self, subsystem, sys_name):
-        DeviceNotFoundError.__init__(self, subsystem, sys_name)
-
-    @property
-    def subsystem(self):
-        """
-        The subsystem that caused this error as string.
-        """
-        return self.args[0]
-
-    @property
-    def sys_name(self):
-        """
-        The sys name that caused this error as string.
-        """
-        return self.args[1]
-
-    def __str__(self):
-        return 'No device {0.sys_name!r} in {0.subsystem!r}'.format(self)
-
-
-class DeviceNotFoundByNumberError(DeviceNotFoundError):
-    """
-    A :exc:`DeviceNotFoundError` indicating, that no :class:`Device` was found
-    for a given device number.
-    """
-
-    def __init__(self, type, number):
-        DeviceNotFoundError.__init__(self, type, number)
-
-    @property
-    def device_type(self):
-        """
-        The device type causing this error as string.  Either ``'char'`` or
-        ``'block'``.
-        """
-        return self.args[0]
-
-    @property
-    def device_number(self):
-        """
-        The device number causing this error as integer.
-        """
-        return self.args[1]
-
-    def __str__(self):
-        return ('No {0.device_type} device with number '
-                '{0.device_number}'.format(self))
-
-
-class DeviceNotFoundInEnvironmentError(DeviceNotFoundError):
-    """
-    A :exc:`DeviceNotFoundError` indicating, that no :class:`Device` could
-    be constructed from the process environment.
-    """
-
-    def __str__(self):
-        return 'No device found in environment'
-
+# pylint: disable=too-many-lines
 
 class Device(Mapping):
+    # pylint: disable=too-many-public-methods
     """
     A single device with attached attributes and properties.
 
@@ -257,7 +168,7 @@ class Device(Mapping):
         return cls(context, device)
 
     @classmethod
-    def from_device_number(cls, context, type, number):
+    def from_device_number(cls, context, typ, number):
         """
         Create a new device from a device ``number`` with the given device
         ``type``:
@@ -295,13 +206,13 @@ class Device(Mapping):
 
         .. versionadded:: 0.11
         """
-        if type not in ('char', 'block'):
+        if typ not in ('char', 'block'):
             raise ValueError('Invalid type: {0!r}. Must be one of "char" '
-                             'or "block".'.format(type))
+                             'or "block".'.format(typ))
         device = context._libudev.udev_device_new_from_devnum(
-            context, ensure_byte_string(type[0]), number)
+            context, ensure_byte_string(typ[0]), number)
         if not device:
-            raise DeviceNotFoundByNumberError(type, number)
+            raise DeviceNotFoundByNumberError(typ, number)
         return cls(context, device)
 
     @classmethod
@@ -816,11 +727,11 @@ class Device(Mapping):
         properties = self._libudev.udev_device_get_properties_list_entry(self)
         return sum(1 for _ in udev_list_iterate(self._libudev, properties))
 
-    def __getitem__(self, property):
+    def __getitem__(self, prop):
         """
-        Get the given ``property`` from this device.
+        Get the given property from this device.
 
-        ``property`` is a unicode or byte string containing the name of the
+        ``prop`` is a unicode or byte string containing the name of the
         property.
 
         Return the property value as unicode string, or raise a
@@ -828,16 +739,16 @@ class Device(Mapping):
         for this device.
         """
         value = self._libudev.udev_device_get_property_value(
-            self, ensure_byte_string(property))
+            self, ensure_byte_string(prop))
         if value is None:
-            raise KeyError(property)
+            raise KeyError(prop)
         return ensure_unicode_string(value)
 
-    def asint(self, property):
+    def asint(self, prop):
         """
-        Get the given ``property`` from this device as integer.
+        Get the given property from this device as integer.
 
-        ``property`` is a unicode or byte string containing the name of the
+        ``prop`` is a unicode or byte string containing the name of the
         property.
 
         Return the property value as integer. Raise a
@@ -845,17 +756,17 @@ class Device(Mapping):
         for this device, or a :exc:`~exceptions.ValueError`, if the property
         value cannot be converted to an integer.
         """
-        return int(self[property])
+        return int(self[prop])
 
-    def asbool(self, property):
+    def asbool(self, prop):
         """
-        Get the given ``property`` from this device as boolean.
+        Get the given property from this device as boolean.
 
         A boolean property has either a value of ``'1'`` or of ``'0'``,
         where ``'1'`` stands for ``True``, and ``'0'`` for ``False``.  Any
         other value causes a :exc:`~exceptions.ValueError` to be raised.
 
-        ``property`` is a unicode or byte string containing the name of the
+        ``prop`` is a unicode or byte string containing the name of the
         property.
 
         Return ``True``, if the property value is ``'1'`` and ``False``, if
@@ -863,7 +774,7 @@ class Device(Mapping):
         :exc:`~exceptions.ValueError`.  Raise a :exc:`~exceptions.KeyError`,
         if the given property is not defined for this device.
         """
-        return string_to_bool(self[property])
+        return string_to_bool(self[prop])
 
     def __hash__(self):
         return hash(self.device_path)
@@ -892,50 +803,6 @@ class Device(Mapping):
     def __ge__(self, other):
         raise TypeError('Device not orderable')
 
-
-class Tags(Iterable, Container):
-    """
-    A iterable over :class:`Device` tags.
-
-    Subclasses the ``Container`` and the ``Iterable`` ABC.
-    """
-
-    def __init__(self, device):
-        self.device = device
-
-    def _has_tag(self, tag):
-        if hasattr(self._libudev, 'udev_device_has_tag'):
-            return bool(self._libudev.udev_device_has_tag(
-                self.device, ensure_byte_string(tag)))
-        else:
-            return any(t == tag for t in self)
-
-    @property
-    def _libudev(self):
-        return self.device._libudev
-
-    def __contains__(self, tag):
-        """
-        Check for existence of ``tag``.
-
-        ``tag`` is a tag as unicode string.
-
-        Return ``True``, if ``tag`` is attached to the device, ``False``
-        otherwise.
-        """
-        return self._has_tag(tag)
-
-    def __iter__(self):
-        """
-        Iterate over all tags.
-
-        Yield each tag as unicode string.
-        """
-        tags = self._libudev.udev_device_get_tags_list_entry(self.device)
-        for tag, _ in udev_list_iterate(self._libudev, tags):
-            yield ensure_unicode_string(tag)
-
-
 def _is_attribute_file(filepath):
     """
     Check, if ``filepath`` points to a valid udev attribute filename.
@@ -951,7 +818,6 @@ def _is_attribute_file(filepath):
     return not (filename.startswith('.') or
                 filename in ('dev', 'uevent') or
                 os.path.islink(filepath))
-
 
 class Attributes(Mapping):
     """
@@ -971,6 +837,9 @@ class Attributes(Mapping):
         self._libudev = device._libudev
 
     def _get_attributes(self):
+        """
+        Yields attributes of device.
+        """
         if hasattr(self._libudev, 'udev_device_get_sysattr_list_entry'):
             attrs = self._libudev.udev_device_get_sysattr_list_entry(
                 self.device)
@@ -1067,3 +936,53 @@ class Attributes(Mapping):
         if the given attribute is not defined for this device.
         """
         return string_to_bool(self.asstring(attribute))
+
+class Tags(Iterable, Container):
+    """
+    A iterable over :class:`Device` tags.
+
+    Subclasses the ``Container`` and the ``Iterable`` ABC.
+    """
+
+    # pylint: disable=too-few-public-methods
+
+    def __init__(self, device):
+        self.device = device
+
+    def _has_tag(self, tag):
+        """
+            Whether ``tag`` exists.
+
+            :param tag: unicode string with name of tag
+            :rtype: bool
+        """
+        if hasattr(self._libudev, 'udev_device_has_tag'):
+            return bool(self._libudev.udev_device_has_tag(
+                self.device, ensure_byte_string(tag)))
+        else:
+            return any(t == tag for t in self)
+
+    @property
+    def _libudev(self):
+        return self.device._libudev
+
+    def __contains__(self, tag):
+        """
+        Check for existence of ``tag``.
+
+        ``tag`` is a tag as unicode string.
+
+        Return ``True``, if ``tag`` is attached to the device, ``False``
+        otherwise.
+        """
+        return self._has_tag(tag)
+
+    def __iter__(self):
+        """
+        Iterate over all tags.
+
+        Yield each tag as unicode string.
+        """
+        tags = self._libudev.udev_device_get_tags_list_entry(self.device)
+        for tag, _ in udev_list_iterate(self._libudev, tags):
+            yield ensure_unicode_string(tag)
